Refactoring for better tesability
